pyproject.toml
[project]
name = "social"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "colorlog>=6.9.0",
    "cryptography>=45.0.7",
    "deprecated>=1.2.18",
    "dj-rest-auth[social,with-social]>=7.0.1",
    "django>=5.2.6",
    "django-allauth>=65.11.2",
    "django-stubs>=5.2.2",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "djangorestframework-stubs>=3.16.2",
    "dotenv>=0.9.9",
    "pillow>=11.3.0",
    "psycopg2-binary>=2.9.10",
    "psycopg[binary]>=3.2.10",
    "requests>=2.32.5",
]

[tool.poe.tasks]
make = "./src/manage.py makemigrations"
migrate = "./src/manage.py migrate"
run = "./src/manage.py runserver"
shell = "./src/manage.py shell"

[dependency-groups]
dev = [
    "poethepoet>=0.37.0",
]

